Added delayed callbacks for body and trailers to DelayedCache and moved it to common.h
